Rename file `activeEditor` => `activeTextEditor`
