+++ src/ide/activeTextEditor.ts
