+++ src/test/suite/editNewCell.test.ts
@@ -5,7 +5,7 @@ import { getCellIndex } from "../../util/notebook";
 import { openNewNotebookEditor } from "../openNewEditor";
 import { getPlainNotebookContents } from "../util/notebook";
 import { sleepWithBackoff, standardSuiteSetup } from "./standardSuiteSetup";
-import { getActiveTextEditor } from "../../ide/activeEditor";
+import { getActiveTextEditor } from "../../ide/activeTextEditor";
 
 // Check that setSelection is able to focus the correct cell
 suite("Edit new cell", async function () {
